+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JX9T0EBM2SWNS1ZPS323ZZGB"
+        "01JXA222QNVXZ7H8NW962BVBM9"
     )
     r"""The id of the resource"""